# Do less allocation in JSDoc parsing