--- a/internal/core/pool.go
+++ b/internal/core/pool.go
@@ -49,6 +49,15 @@ func (p *Pool[T]) NewSlice1(t T) []T {
 	return slice
 }
 
+func (p *Pool[T]) Clone(t []T) []T {
+	if len(t) == 0 {
+		return nil
+	}
+	slice := p.NewSlice(len(t))
+	copy(slice, t)
+	return slice
+}
+
 func nextPoolSize(size int) int {
 	// This compiles down branch-free.
 	size = max(size, 1)