One of the largest allocators when parsing code for the editor is JSDoc parsing. By deferring JSDoc part joining until `Text()` is called, we can save a bunch of memory allocations, and make parsing out the DOM types ~5% faster.

```
goos: linux
goarch: amd64
pkg: github.com/microsoft/typescript-go/internal/parser
cpu: Intel(R) Core(TM) i9-10900K CPU @ 3.70GHz
                                                             │   old.txt   │              new.txt               │
                                                             │   sec/op    │   sec/op     vs base               │
Parse/empty.ts/tsc-20                                          527.5n ± 4%   523.9n ± 1%       ~ (p=0.072 n=10)
Parse/empty.ts/server-20                                       524.1n ± 1%   527.3n ± 1%       ~ (p=0.138 n=10)
Parse/checker.ts/tsc-20                                        48.66m ± 2%   47.70m ± 1%  -1.97% (p=0.002 n=10)
Parse/checker.ts/server-20                                     48.88m ± 4%   49.15m ± 1%       ~ (p=0.912 n=10)
Parse/dom.generated.d.ts/tsc-20                                16.39m ± 2%   16.11m ± 1%  -1.69% (p=0.023 n=10)
Parse/dom.generated.d.ts/server-20                             24.56m ± 2%   23.54m ± 3%  -4.13% (p=0.000 n=10)
Parse/Herebyfile.mjs/tsc-20                                    884.3µ ± 1%   881.8µ ± 1%       ~ (p=0.143 n=10)
Parse/Herebyfile.mjs/server-20                                 886.8µ ± 3%   880.6µ ± 2%       ~ (p=0.165 n=10)
Parse/jsxComplexSignatureHasApplicabilityError.tsx/tsc-20      267.4µ ± 2%   265.2µ ± 2%  -0.80% (p=0.015 n=10)
Parse/jsxComplexSignatureHasApplicabilityError.tsx/server-20   419.1µ ± 2%   403.0µ ± 3%  -3.83% (p=0.000 n=10)
geomean                                                        686.5µ        677.5µ       -1.31%
```
```
                                                             │   old.txt    │                new.txt                │
                                                             │     B/op     │     B/op      vs base                 │
Parse/empty.ts/tsc-20                                            881.0 ± 0%     881.0 ± 0%       ~ (p=1.000 n=10) ¹
Parse/empty.ts/server-20                                         881.0 ± 0%     881.0 ± 0%       ~ (p=1.000 n=10) ¹
Parse/checker.ts/tsc-20                                        25.67Mi ± 0%   25.66Mi ± 0%       ~ (p=0.684 n=10)
Parse/checker.ts/server-20                                     25.84Mi ± 0%   25.84Mi ± 0%  +0.02% (p=0.023 n=10)
Parse/dom.generated.d.ts/tsc-20                                9.154Mi ± 0%   9.153Mi ± 0%       ~ (p=0.971 n=10)
Parse/dom.generated.d.ts/server-20                             11.51Mi ± 0%   11.22Mi ± 0%  -2.52% (p=0.000 n=10)
Parse/Herebyfile.mjs/tsc-20                                    447.6Ki ± 0%   448.3Ki ± 0%  +0.15% (p=0.000 n=10)
Parse/Herebyfile.mjs/server-20                                 447.6Ki ± 0%   448.3Ki ± 0%  +0.15% (p=0.000 n=10)
Parse/jsxComplexSignatureHasApplicabilityError.tsx/tsc-20      162.3Ki ± 0%   162.2Ki ± 0%  -0.02% (p=0.000 n=10)
Parse/jsxComplexSignatureHasApplicabilityError.tsx/server-20   239.1Ki ± 0%   242.1Ki ± 0%  +1.28% (p=0.000 n=10)
geomean                                                        461.9Ki        461.4Ki       -0.10%
¹ all samples are equal
```
```
                                                             │   old.txt   │                new.txt                │
                                                             │  allocs/op  │  allocs/op   vs base                  │
Parse/empty.ts/tsc-20                                           4.000 ± 0%    4.000 ± 0%        ~ (p=1.000 n=10) ¹
Parse/empty.ts/server-20                                        4.000 ± 0%    4.000 ± 0%        ~ (p=1.000 n=10) ¹
Parse/checker.ts/tsc-20                                        29.18k ± 0%   29.18k ± 0%   -0.02% (p=0.000 n=10)
Parse/checker.ts/server-20                                     29.80k ± 0%   29.50k ± 0%   -1.00% (p=0.000 n=10)
Parse/dom.generated.d.ts/tsc-20                                23.01k ± 0%   23.01k ± 0%        ~ (p=0.582 n=10)
Parse/dom.generated.d.ts/server-20                             30.46k ± 0%   25.02k ± 0%  -17.86% (p=0.000 n=10)
Parse/Herebyfile.mjs/tsc-20                                    1.880k ± 0%   1.887k ± 0%   +0.37% (p=0.000 n=10)
Parse/Herebyfile.mjs/server-20                                 1.880k ± 0%   1.887k ± 0%   +0.37% (p=0.000 n=10)
Parse/jsxComplexSignatureHasApplicabilityError.tsx/tsc-20       331.0 ± 0%    331.0 ± 0%        ~ (p=1.000 n=10) ¹
Parse/jsxComplexSignatureHasApplicabilityError.tsx/server-20    597.0 ± 0%    457.0 ± 0%  -23.45% (p=0.000 n=10)
geomean                                                        1.211k        1.156k        -4.56%
¹ all samples are equal
```